+++ src/main/js/types/PullRequest.ts
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-import { Collection, DisplayedUser, Links } from "@scm-manager/ui-types";
+import { Collection, DisplayedUser, Links, Embedded, HalRepresentation } from "@scm-manager/ui-types";
 
 export type Reviewer = DisplayedUser & {
   approved: boolean;
@@ -34,7 +34,7 @@ export type BasicPullRequest = {
   title: string;
 };
 
-export type PullRequest = BasicPullRequest & {
+export type PullRequest = BasicPullRequest & HalRepresentation & {
   description?: string;
   author: DisplayedUser;
   reviser?: DisplayedUser;
@@ -42,7 +42,6 @@ export type PullRequest = BasicPullRequest & {
   creationDate: string;
   reviewer: Reviewer[];
   status: string;
-  _links: Links;
   tasks: Tasks;
   sourceRevision: string;
   targetRevision: string;